requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
typing_extensions==4.15.0
Werkzeug==3.1.5
WTForms==3.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
